Move builder mixins to package
