+++ src/main/kotlin/io/github/freya022/botcommands/internal/components/builder/BuilderInstanceHolder.kt
@@ -4,8 +4,3 @@ package io.github.freya022.botcommands.internal.components.builder
 internal interface BuilderInstanceHolder<T : Any> {
     val instance: T
 }
-
-internal inline fun <T : Any> BuilderInstanceHolder<T>.applyInstance(block: () -> Unit): T {
-    block() // No need to give the instance
-    return instance
-}
