+++ src/main/kotlin/io/github/freya022/botcommands/internal/components/builder/group/EphemeralComponentGroupBuilderImpl.kt
@@ -3,9 +3,9 @@ package io.github.freya022.botcommands.internal.components.builder.group
 import io.github.freya022.botcommands.api.components.IGroupHolder
 import io.github.freya022.botcommands.api.components.builder.group.EphemeralComponentGroupBuilder
 import io.github.freya022.botcommands.internal.components.LifetimeType
-import io.github.freya022.botcommands.internal.components.builder.EphemeralTimeoutableComponentImpl
-import io.github.freya022.botcommands.internal.components.builder.IEphemeralTimeoutableComponentMixin
 import io.github.freya022.botcommands.internal.components.builder.InstanceRetriever
+import io.github.freya022.botcommands.internal.components.builder.mixin.IEphemeralTimeoutableComponentMixin
+import io.github.freya022.botcommands.internal.components.builder.mixin.impl.EphemeralTimeoutableComponentImpl
 import io.github.freya022.botcommands.internal.components.controller.ComponentController
 
 internal class EphemeralComponentGroupBuilderImpl internal constructor(
