+++ src/main/kotlin/io/github/freya022/botcommands/internal/components/builder/mixin/BaseComponentBuilderMixin.kt
@@ -1,4 +1,4 @@
-package io.github.freya022.botcommands.internal.components.builder
+package io.github.freya022.botcommands.internal.components.builder.mixin
 
 import io.github.freya022.botcommands.api.components.builder.BaseComponentBuilder
