+++ src/main/kotlin/io/github/freya022/botcommands/internal/components/builder/mixin/IActionableComponentMixin.kt
@@ -1,9 +1,10 @@
-package io.github.freya022.botcommands.internal.components.builder
+package io.github.freya022.botcommands.internal.components.builder.mixin
 
 import io.github.freya022.botcommands.api.components.builder.IActionableComponent
 import io.github.freya022.botcommands.api.components.builder.IEphemeralActionableComponent
 import io.github.freya022.botcommands.api.components.builder.IPersistentActionableComponent
 import io.github.freya022.botcommands.api.components.ratelimit.ComponentRateLimitReference
+import io.github.freya022.botcommands.internal.components.builder.BuilderInstanceHolder
 import io.github.freya022.botcommands.internal.components.handler.ComponentHandler
 import net.dv8tion.jda.api.events.interaction.component.GenericComponentInteractionCreateEvent
