+++ src/main/kotlin/io/github/freya022/botcommands/internal/components/builder/mixin/ITimeoutableComponentMixin.kt
@@ -1,8 +1,9 @@
-package io.github.freya022.botcommands.internal.components.builder
+package io.github.freya022.botcommands.internal.components.builder.mixin
 
 import io.github.freya022.botcommands.api.components.builder.IEphemeralTimeoutableComponent
 import io.github.freya022.botcommands.api.components.builder.IPersistentTimeoutableComponent
 import io.github.freya022.botcommands.api.components.builder.ITimeoutableComponent
+import io.github.freya022.botcommands.internal.components.builder.BuilderInstanceHolder
 import io.github.freya022.botcommands.internal.components.data.timeout.ComponentTimeout
 import kotlin.time.Duration
