+++ src/main/kotlin/io/github/freya022/botcommands/internal/components/builder/mixin/IUniqueComponentMixin.kt
@@ -1,6 +1,7 @@
-package io.github.freya022.botcommands.internal.components.builder
+package io.github.freya022.botcommands.internal.components.builder.mixin
 
 import io.github.freya022.botcommands.api.components.builder.IUniqueComponent
+import io.github.freya022.botcommands.internal.components.builder.BuilderInstanceHolder
 
 internal interface IUniqueComponentMixin<T : IUniqueComponent<T>> : IUniqueComponent<T>,
                                                                     BuilderInstanceHolder<T>
