+++ src/main/kotlin/io/github/freya022/botcommands/internal/components/builder/mixin/impl/PersistentActionableComponentImpl.kt
@@ -1,11 +1,14 @@
 @file:Suppress("DEPRECATION")
 
-package io.github.freya022.botcommands.internal.components.builder
+package io.github.freya022.botcommands.internal.components.builder.mixin.impl
 
 import io.github.freya022.botcommands.api.ReceiverConsumer
 import io.github.freya022.botcommands.api.components.builder.IPersistentActionableComponent
 import io.github.freya022.botcommands.api.components.builder.PersistentHandlerBuilder
 import io.github.freya022.botcommands.api.core.BContext
+import io.github.freya022.botcommands.internal.components.builder.AbstractActionableComponent
+import io.github.freya022.botcommands.internal.components.builder.InstanceRetriever
+import io.github.freya022.botcommands.internal.components.builder.mixin.IPersistentActionableComponentMixin
 import io.github.freya022.botcommands.internal.components.handler.PersistentHandler
 
 internal class PersistentActionableComponentImpl<T : IPersistentActionableComponent<T>> internal constructor(
