+++ src/main/kotlin/io/github/freya022/botcommands/internal/components/builder/mixin/impl/PersistentTimeoutableComponentImpl.kt
@@ -1,6 +1,9 @@
-package io.github.freya022.botcommands.internal.components.builder
+package io.github.freya022.botcommands.internal.components.builder.mixin.impl
 
 import io.github.freya022.botcommands.api.components.builder.IPersistentTimeoutableComponent
+import io.github.freya022.botcommands.internal.components.builder.BuilderInstanceHolderImpl
+import io.github.freya022.botcommands.internal.components.builder.InstanceRetriever
+import io.github.freya022.botcommands.internal.components.builder.mixin.IPersistentTimeoutableComponentMixin
 import io.github.freya022.botcommands.internal.components.data.timeout.PersistentTimeout
 import io.github.freya022.botcommands.internal.utils.Checks
 import kotlin.time.Duration
