+++ src/main/kotlin/io/github/freya022/botcommands/internal/components/repositories/ComponentRepository.kt
@@ -8,10 +8,10 @@ import io.github.freya022.botcommands.api.core.db.*
 import io.github.freya022.botcommands.api.core.service.annotations.BService
 import io.github.freya022.botcommands.internal.components.ComponentType
 import io.github.freya022.botcommands.internal.components.LifetimeType
-import io.github.freya022.botcommands.internal.components.builder.BaseComponentBuilderMixin
-import io.github.freya022.botcommands.internal.components.builder.IComponentBuilderMixin
-import io.github.freya022.botcommands.internal.components.builder.ITimeoutableComponentMixin
 import io.github.freya022.botcommands.internal.components.builder.group.AbstractComponentGroupBuilder
+import io.github.freya022.botcommands.internal.components.builder.mixin.BaseComponentBuilderMixin
+import io.github.freya022.botcommands.internal.components.builder.mixin.IComponentBuilderMixin
+import io.github.freya022.botcommands.internal.components.builder.mixin.ITimeoutableComponentMixin
 import io.github.freya022.botcommands.internal.components.controller.ComponentFilters
 import io.github.freya022.botcommands.internal.components.data.ComponentData
 import io.github.freya022.botcommands.internal.components.data.ComponentGroupData
